v0.0.3 - transforming old version into add-on
