+++ MapDriveInator/build.gradle
@@ -1,47 +1,76 @@
-//MapDriveInator
+//groovyConsole
+
+
 
 apply plugin: 'groovy'
 
 targetCompatibility='1.8'
 
+// 1. I have to define which freeplane version is going to be used to look at the libreries (freeplane and groovy)
+def FPversion = 'v10_USS'
+def freeplaneDirs = [
+        instaled     : "C:/Program Files/Freeplane",
+        v10_portableApps : "C:/PortableApps/FreeplanePortable/App/Freeplane",
+        v9_15         : "C:/Freeplane/freeplane-1.9.15-pre01",
+        v10_1         : "C:/Freeplane/freeplane-1.10.2",
+        v10_USS       : "C:/Users/eduardo.frohlich/PortableApps/FreeplanePortable/App/Freeplane"
+    ]
+
+def freeplaneDir = freeplaneDirs[FPversion]
+
+def groovyGroupId = FPversion.startsWith('v10')?'org.apache.groovy':'org.codehaus.groovy'
+def groovyVersion = FPversion.startsWith('v10')?'4.0.1':'3.0.8'
+
+
 libsDirName = "${rootDir}/lib"
-clean.doFirst { delete "${rootDir}/lib" }
-//clean.doFirst { delete "${rootDir}/lib/WikdShellExtension.jar" } //para eliminar sólo una de las .jar
+clean.doFirst { delete "${rootDir}/lib" } // para eliminar todas las .jar
+//clean.doFirst { delete "${rootDir}/lib/Tutorial-o-Matic.jar" } //para eliminar sólo una de las .jar y que no me borrara markedj ni jsoup
 
 repositories {
     mavenCentral()
     flatDir name: 'localGitDepsRepository',
     dirs: [
-        "C:/Program Files/Freeplane",
-        "C:/Program Files/Freeplane/core/org.freeplane.core/lib",
-        "C:/Program Files/Freeplane/plugins/org.freeplane.plugin.script/lib"
-        //, "${rootDir}/lib" //así lo usé en wikdshellext, porque ahí puse WikdShell.jar
+        "${freeplaneDir}",
+        "${freeplaneDir}/core/org.freeplane.core/lib",
+        "${freeplaneDir}/plugins/org.freeplane.plugin.script/lib",
+        // "${freeplaneDir}/plugins/org.freeplane.plugin.markdown/lib", // para incluir markedj.jar
+        "${rootDir}/lib" //ahí puse markedj y jsoup
     ]
-
-    // Así para cuando Freeplane está en Portable Apps
-        // dirs: [
-        // "C:/PortableApps/FreeplanePortable/App/Freeplane",
-        // "C:/PortableApps/FreeplanePortable/App/Freeplane/core/org.freeplane.core/lib",
-        // "C:/PortableApps/FreeplanePortable/App/Freeplane/plugins/org.freeplane.plugin.script/lib",
-        // "${rootDir}/lib" //así lo usé en wikdshellext, porque ahí puse WikdShell.jar
-        // ]
 }
 
+
+// I had this error:
+    // Could not find method compile() for arguments [{name=freeplanelauncher}] on object of type org.gradle.api.internal.artifacts.dsl.dependencies.DefaultDependencyHandler
+// answer:
+    // https://stackoverflow.com/questions/69733508/couldnt-find-compile-for-arguments-directory-libs-on-object-of-type-org-g
+
 dependencies {
-    compile name: 'freeplanelauncher'
-    compile name: 'freeplaneviewer'
-    compile group: 'org.freeplane.core', name: 'freeplaneeditor', version:"+"
-    compile group: 'org.freeplane.script', name: 'plugin', version: '+'
-    // compile name: 'WikdShell' //así lo usé en wikdshellext.
-    compile ('org.codehaus.groovy:groovy-all:3.0.1') {
-        exclude group:'org.codehaus.groovy', module:'groovy-test'
-        exclude group:'org.codehaus.groovy', module:'groovy-test-junit5'
-        exclude group:'org.codehaus.groovy', module:'groovy-testng'
-        exclude group:'org.codehaus.groovy', module:'groovy-ant'
-        exclude group:'org.codehaus.groovy', module:'groovy-docgenerator'
-        exclude group:'org.codehaus.groovy', module:'groovy-groovydoc'
-        exclude group:'org.codehaus.groovy', module:'groovy-cli-commons'
-        exclude group:'org.codehaus.groovy', module:'groovy-cli-picocli'
+    implementation name : 'freeplanelauncher'
+    implementation name : 'freeplaneviewer'
+    implementation group: 'org.freeplane.core',   name: 'freeplaneeditor', version:"+"
+    implementation group: 'org.freeplane.script', name: 'plugin', version: '+'
+    if(FPversion.startsWith('v10')){
+        implementation platform ("${groovyGroupId}:groovy-all:${groovyVersion}".toString()) {
+            exclude group:groovyGroupId, module:'groovy-test'
+            exclude group:groovyGroupId, module:'groovy-test-junit5'
+            exclude group:groovyGroupId, module:'groovy-testng'
+            exclude group:groovyGroupId, module:'groovy-ant'
+            exclude group:groovyGroupId, module:'groovy-docgenerator'
+            exclude group:groovyGroupId, module:'groovy-groovydoc'
+            exclude group:groovyGroupId, module:'groovy-cli-commons'
+            exclude group:groovyGroupId, module:'groovy-cli-picocli'
+        }
+    } else {
+        implementation ("${groovyGroupId}:groovy-all:${groovyVersion}".toString()) {
+            exclude group:groovyGroupId, module:'groovy-test'
+            exclude group:groovyGroupId, module:'groovy-test-junit5'
+            exclude group:groovyGroupId, module:'groovy-testng'
+            exclude group:groovyGroupId, module:'groovy-ant'
+            exclude group:groovyGroupId, module:'groovy-docgenerator'
+            exclude group:groovyGroupId, module:'groovy-groovydoc'
+            exclude group:groovyGroupId, module:'groovy-cli-commons'
+            exclude group:groovyGroupId, module:'groovy-cli-picocli'
+        }
     }
-    compile 'org.codehaus.groovy:groovy-dateutil:3.0.1'
+    implementation "${groovyGroupId}:groovy-dateutil:${groovyVersion}".toString()
 }
