+++ MapDriveInator/history.md
@@ -1,0 +1,5 @@
+# History
+
+## v0.0.3
+
+* old script version packed as an Add-on
