+++ MapDriveInator/scripts/groupChildnodesBy.groovy
