+++ MapDriveInator/scripts/unmarkMovedOrRenamedNodes.groovy
@@ -1,0 +1,6 @@
+
+import MDI
+
+node.find{MDI.isMoved(it)}.each{MDI.markAsMoved(it,false)}
+
+
