+++ MapDriveInator/version.properties
@@ -1,3 +1,4 @@
-version=v0.0.2
-downloadUrl=https://github.com/EdoFro/Freeplane_Map-Drive-Inator/releases/download/v0.0.2/MapDriveInator-v0.0.2.addon.mm
+version=v0.0.3
+downloadUrl=https://github.com/EdoFro/Freeplane_Map-Drive-Inator/releases/download/v0.0.3/MapDriveInator-v0.0.3.addon.mm
+changelogUrl=https://github.com/EdoFro/Freeplane_Map-Drive-Inator/history.txt
 freeplaneVersionFrom=v1.9.0
