+++ MapDriveInatorV2/version.properties
@@ -1,0 +1,3 @@
+version=v0.0.2
+downloadUrl=https://github.com/EdoFro/Freeplane_Map-Drive-Inator/releases/download/v0.0.2/MapDriveInator-v0.0.2.addon.mm
+freeplaneVersionFrom=v1.9.0
